feat(admin): split test & admin
